requirements.txt
flask==3.0.2
gunicorn==21.2.0
python-dotenv==1.0.1
chromadb==0.5.3
rank-bm25==0.2.2
sentence-transformers==2.7.0
numpy==1.26.4
groq==0.9.0
edge-tts==6.1.12
livekit==0.11.1
livekit-api==1.1.0
livekit-agents==1.4.3
